requirements.txt
streamlit
openai
python-pptx
requests
pillow
comtypes; platform_system == "Windows"
